backend/requirements.txt
blinker==1.9.0
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.3.0
colorama==0.4.6
filelock==3.19.1
Flask==3.1.2
Flask-Mail
flask-cors
flask-bcrypt
flask-jwt-extended
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
python-dotenv
python-dateutil==2.9.0.post0
python-whois==0.9.5
whoisit
pymongo
requests==2.32.5
requests-file==2.1.0
six==1.17.0
tabulate==0.9.0
tldextract==5.3.0
urllib3>=2.6.3
Werkzeug==3.1.3

joblib==1.4.2
numpy==1.26.4
scikit-learn==1.5.2

celery
redis
flask-limiter[redis]>=3.5.0
